Add board node and heuristics files
This is to start exploring and experimenting with different search
algorithms, without having to change the underlying Board Model. This
way, all the algorithms will interact with is the node, which will act
as the bridge between algorithms and the actual model. The heuristics are
merely a scaffold to start getting ideas on what is necessary or not